Move AlphaVantage's data response classes to its own package
