+++ app/src/main/java/com/farmingdale/stockscreener/model/remote/alphavantageResponses/QuoteDataResponse.kt
@@ -1,4 +1,4 @@
-package com.farmingdale.stockscreener.model.remote
+package com.farmingdale.stockscreener.model.remote.alphavantageResponses
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
